Refactor sinescroll and add newer cracktro screen
Sinescroll is no longer a flag, and is now it's own separate thing.
Using `V_SetAmigaEffect` and co, any patch drawn after that will be
drawn with the sinescroll wave effect. Use `V_ResetAmigaCounter` to
reset the counter for it. `V_ClearAmigaEffect` will not apply the
effect to future patch draws entirely.

Not compatible with clip rects.